+++ changelog.md
@@ -8,6 +8,45 @@
 * Upgraded project reactor to 3.4.10
 * Upgrading to cloudevents sdk 2.2.0
 * Minor tweak in ApplicationService extension function for Kotlin so that it no longer converts the Java stream into a temporary Kotlin sequence before converting it to a List
+* Allow configuring whether transactional reads should be enabled or disabled for all MongoDB event stores.
+  This is an advanced feature, and you almost always want to have it enabled. There are two reasons for disabling it:
+  1. There's a bug/limitation on Atlas free tier clusters which yields an exception when reading large number of events in a stream in a transaction.
+     To workaround this you could disable transactional reads. The exception takes this form:
+     ```
+     java.lang.IllegalStateException: state should be: open
+     at com.mongodb.assertions.Assertions.isTrue(Assertions.java:79)
+     at com.mongodb.internal.session.BaseClientSessionImpl.getServerSession(BaseClientSessionImpl.java:101)
+     at com.mongodb.internal.session.ClientSessionContext.getSessionId(ClientSessionContext.java:44)
+     at com.mongodb.internal.connection.ClusterClockAdvancingSessionContext.getSessionId(ClusterClockAdvancingSessionContext.java:46)
+     at com.mongodb.internal.connection.CommandMessage.getExtraElements(CommandMessage.java:265)
+     at com.mongodb.internal.connection.CommandMessage.encodeMessageBodyWithMetadata(CommandMessage.java:155)
+     at com.mongodb.internal.connection.RequestMessage.encode(RequestMessage.java:138)
+     at com.mongodb.internal.connection.CommandMessage.encode(CommandMessage.java:59)
+     at com.mongodb.internal.connection.InternalStreamConnection.sendAndReceive(InternalStreamConnection.java:268)
+     at com.mongodb.internal.connection.UsageTrackingInternalConnection.sendAndReceive(UsageTrackingInternalConnection.java:100)
+     at com.mongodb.internal.connection.DefaultConnectionPool$PooledConnection.sendAndReceive(DefaultConnectionPool.java:490)
+     at com.mongodb.internal.connection.CommandProtocolImpl.execute(CommandProtocolImpl.java:71)
+     at com.mongodb.internal.connection.DefaultServer$DefaultServerProtocolExecutor.execute(DefaultServer.java:253)
+     at com.mongodb.internal.connection.DefaultServerConnection.executeProtocol(DefaultServerConnection.java:202)
+     at com.mongodb.internal.connection.DefaultServerConnection.command(DefaultServerConnection.java:118)
+     at com.mongodb.internal.connection.DefaultServerConnection.command(DefaultServerConnection.java:110)
+     at com.mongodb.internal.operation.QueryBatchCursor.getMore(QueryBatchCursor.java:268)
+     at com.mongodb.internal.operation.QueryBatchCursor.hasNext(QueryBatchCursor.java:141)
+     at com.mongodb.client.internal.MongoBatchCursorAdapter.hasNext(MongoBatchCursorAdapter.java:54)
+     at java.base/java.util.Iterator.forEachRemaining(Iterator.java:132)
+     at java.base/java.util.Spliterators$IteratorSpliterator.forEachRemaining(Spliterators.java:1801)
+     at java.base/java.util.stream.AbstractPipeline.copyInto(AbstractPipeline.java:484)
+     at java.base/java.util.stream.AbstractPipeline.wrapAndCopyInto(AbstractPipeline.java:474)
+     at java.base/java.util.stream.ReduceOps$ReduceOp.evaluateSequential(ReduceOps.java:913)
+     at java.base/java.util.stream.AbstractPipeline.evaluate(AbstractPipeline.java:234)
+     ```
+     It's possible that this would work if you enable "no cursor timeout" on the query, but this is not allowed on Atlas free tier.
+  2. You're set back by the performance penalty of transactions and are willing to sacrifice read consistency
+  
+  If you disable transactional reads, you _may_ end up with a mismatch between the version number in the `EventStream` and
+  the last event returned from the event stream. This is because Occurrent does two reads to MongoDB when reading an event stream. First it finds the current version number of the stream (A),
+  and secondly it queries for all events (B). If you disable transactional reads, then another thread might have written more events before the call to B has been made. Thus, the version number
+  received from query A might be stale. This may or may not be a problem for your domain, but it's generally recommended having transactional reads enabled.
 
 ## Changelog 0.11.0 (2021-08-13)